WIP connect mint button
