+++ app/helpers/getTokenBalance.tsx
@@ -1,8 +1,8 @@
 import { Contract } from "ethers";
 
-const getTokenBalance = async (erc20: Contract, address: string) => {
-	const coinBalance = await erc20.balanceOf(address);
-	return coinBalance.toNumber();
+const getTokenBalance = async (erc721: Contract, address: string) => {
+  const coinBalance = await erc721.balanceOf(address);
+  return coinBalance.toNumber();
 };
 
 export default getTokenBalance;
